+++ scripts/geojson/GeoJSONProps.ts
@@ -1,21 +1,20 @@
-import { GeoJSONOptions, BaseIcon } from "leaflet";
-import { MapObservableFactory } from "../layer/MapContext";
+import { GeoJSONOptions, Icon } from "leaflet";;
+import { ObservableLayerProps } from "../layer/ObservableLayer";
 
-type SupportedGeometries = GeoJSON.Point | GeoJSON.Polygon | GeoJSON.MultiPolygon;
+export type SupportedGeometries = GeoJSON.Point | GeoJSON.Polygon | GeoJSON.MultiPolygon;
 
 export type GeoJSONCollection = GeoJSON.FeatureCollection<SupportedGeometries>;
 
 export type GeoJSONFeature = GeoJSON.Feature<SupportedGeometries>;
 
-export type GeoJSONProps = GeoJSONOptions & {
-    observable: MapObservableFactory<GeoJSONCollection>,
-    icon?: (feature: GeoJSONFeature) => BaseIcon,
+export type GeoJSONProps = ObservableLayerProps<GeoJSONCollection> & GeoJSONOptions & {
+    icon?: (feature: GeoJSONFeature) => Icon,
     onMarkerClick?: (feature: GeoJSONFeature) => void,
     popup?: (feature: GeoJSONFeature) => JSX.Element,
-    featureId: (feature: GeoJSONFeature) => string
+    featureId?: (feature: GeoJSONFeature) => string
 };
 
 export type ClusterProps = GeoJSONProps & {
     isCluster: (feature: GeoJSONFeature) => boolean;
-    clusterIcon?: (feature: GeoJSONFeature) => BaseIcon,
+    clusterIcon?: (feature: GeoJSONFeature) => Icon,
 };