+++ test/LayerBinderSpec.ts
@@ -4,7 +4,6 @@ import expect = require("expect.js");
 import {IMock, Mock, Times, It} from "typemoq";
 import {ReplaySubject, Subject, Observable} from "rx";
 import {LatLng, LatLngBounds} from "leaflet";
-import MockLayerView from "./fixtures/MockLayerView";
 import IMapBoundaries from "../scripts/leaflet/IMapBoundaries";
 import ILayerView from "../scripts/layer/ILayerView";
 import LayerBinder from "../scripts/layer/LayerBinder";
@@ -23,7 +22,8 @@ describe("Given a layer binder", () => {
         mapBoundaries = Mock.ofType<IMapBoundaries>();
         mapBoundaries.setup(m => m.boundsChanges()).returns(() => viewChanges);
         data = new ReplaySubject<any>();
-        layerView = Mock.ofType<ILayerView<any, any>>(MockLayerView);
+        layerView = Mock.ofType<ILayerView<any, any>>();
+        layerView.setup(l => l.type).returns(() => "GeoJSON");
         subject = new LayerBinder([layerView.object], mapBoundaries.object);
     });
 
@@ -58,19 +58,31 @@ describe("Given a layer binder", () => {
             data.onNext({markers: []});
             data.onNext({markers: [{id: "8283"}]});
         });
-        it("should reload the source with the new bounding box", () => {
-            subject.bind(context => {
-                if (!context.bounds) return data;
-                return Observable.just(context);
-            }, "GeoJSON", null)[1].subscribe();
-            mapBoundaries.setup(m => m.getBounds()).returns(() => new LatLngBounds(new LatLng(50, 50), new LatLng(60, 80)));
-            mapBoundaries.setup(m => m.getZoom()).returns(() => 12);
-            viewChanges.onNext(null);
+        context("when those changes are enabled", () => {
+            it("should reload the source with the new bounding box", () => {
+                subject.bind(context => {
+                    if (!context.bounds) return data;
+                    return Observable.just(context);
+                }, "GeoJSON", null)[1].subscribe();
+                mapBoundaries.setup(m => m.getBounds()).returns(() => new LatLngBounds(new LatLng(50, 50), new LatLng(60, 80)));
+                mapBoundaries.setup(m => m.getZoom()).returns(() => 12);
+                viewChanges.onNext(null);
 
-            layerView.verify(g => g.update(It.isAny(), It.isValue({
-                bounds: new LatLngBounds(new LatLng(50, 50), new LatLng(60, 80)),
-                zoom: 12
-            }), It.isAny(), null), Times.once());
+                layerView.verify(g => g.update(It.isAny(), It.isValue({
+                    bounds: new LatLngBounds(new LatLng(50, 50), new LatLng(60, 80)),
+                    zoom: 12
+                }), It.isAny(), null), Times.once());
+            });
+        });
+        context("when those changes are disabled", () => {
+            it("should not trigger an update", () => {
+                subject.bind(context => data, "GeoJSON", {
+                    freezeBounds: true
+                })[1].subscribe();
+                viewChanges.onNext(null);
+
+                layerView.verify(g => g.update(It.isAny(), It.isAny(), It.isAny(), It.isAny()), Times.exactly(2));
+            });
         });
     });
 });